Support for DB_READ_REPLICA_HOST
We make use of a new hm-platform db dropin to bridge LudicrousDB and
XRay. Our DB dropin is _always_ used, but only a second DB server is
added if the constant is set.

We also send the different hostnames to XRay for DB queries so we
correctly the get service map showing amount of traffic to the db or the
db replica.

We also default the DB_CHARSET to utfmb4, as `wpdb` has done this for a
while, but LudicrousDB never had that functionality.