Disable tcp responsiveness in ludicrous
Ludicrous _can_ check TCP responsiveness of servers, to chose the one that has a closer round-trip. We don't really need the overhead of this "feature" and would be simpler without it. It also requires the object-cache API is available before the WPDB, which isn't really "correct". WordPress inits the DB _before_ object cache.